appender: `AddWithCallback` method implementation
Added the `AddWithCallback` method to the Appender. This method allows
the user to add a document to the bulk indexer and execute a callback
function when the document is processed.

Processed is defined as:

- The document is successfully indexed.
- The document is not indexed due to a terminal error.

Additionally, adds a new `hot` channel to the appender to:

- Reduce the memory footprint of the appender on low throughput.
- Reduce the meant time to retries on high throughput, since the `hot`
channel pulls are prioritized over the `available` channel pulls.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>